requirements.txt
setuptools>=68.0.0
wheel>=0.41.0
binance-futures-connector>=4.1.0
streamlit>=1.29.0
pandas>=2.1.0
numpy>=1.25.0
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.17.0
schedule>=1.2.0

